Bug: SQlbing did not working, so changed that
using tradtion SQLALCHEMY_DATABASE_URI and remove few uncesesary line in routes.py Product folder

+++ Files/product/routes.py
@@ -9,43 +9,21 @@ product = Blueprint('product', __name__, url_prefix='/product')
 from .utils import add_product_args, resource_fields
 
 @product.get('/')
-@marshal_with(resource_fields)
+# @marshal_with(resource_fields)
 def get_products():
     result=Product.query.all()
     return jsonify(result)
 
 @product.get('/<int:id>')
-@marshal_with(resource_fields)
+# @marshal_with(resource_fields)
 def get_product(id):
     result=db.session.query(Product).filter(Product.product_id==id).first()
     if not result:
-        abort(404, message="Product {} doesn't exist".format(id))
+        return {"message": "No product find"}, 404
     return jsonify(result)
 
-# @product.post('/')
-# @marshal_with(resource_fields)
-# def add_product():
-#     if request.is_json:
-#         args=add_product_args.parse_args()
-#         product_name=args['name']
-#         product_description=args['description']
-#         product_price=args['price']
-#         product_image=args['image']
-#         product_discount=args['discount']
-#         product_qty=args['qty_left']
-#         product_category=args['category']
-#         product_related_products=args['related_products']
-#         product=Product(name=product_name,description=product_description,price=product_price,
-#             image=product_image,discount=product_discount,effective_price=product_price-(product_discount*product_price/100),
-#             qty_left=product_qty,category=product_category,related_products=product_related_products)
-#         db.session.add(product)
-#         db.session.commit()
-#         return product, 201
-
-#     return {"error": "Request must be JSON"}, 415
-
 @product.post('/')
-@marshal_with(resource_fields)
+# @marshal_with(resource_fields)
 def add_product():
     if request.is_json:
         product=Product(name=request.json['name'],description=request.json['description'],price=request.json['price'],
@@ -53,16 +31,18 @@ def add_product():
             qty_left=request.json['qty_left'],category=request.json['category'],related_products=request.json['related_products'])
         db.session.add(product)
         db.session.commit()
-        return jsonify(product), 201
+        return {"message": "Dones"}, 201
 
-    return {"error": "Request must be JSON"}, 415
+    return {"message": "Request must be JSON"}, 415
 
 @product.post("/update/<int:id>")
-@marshal_with(resource_fields)
+# @marshal_with(resource_fields)
 def modify_product(id):
     #modify product details like price, etc.
     if request.is_json:
         product=db.session.query(Product).filter(Product.product_id==id).first()
+        if not product:
+            return {"message": "No product find"}, 404
         args=add_product_args.parse_args()
         
         product.name=args['name']
@@ -82,10 +62,12 @@ def modify_product(id):
     
     
 @product.post("/delete/<int:id>")
-@marshal_with(resource_fields)
+# @marshal_with(resource_fields)
 def delete_product(id):
     #delete specified product
     product=db.session.query(Product).filter(Product.product_id==id).first()
+    if not product:
+        return {"message": "No product find"}, 404
     db.session.delete(product)
     db.session.commit()
-    return jsonify(product), 201
+    return {"message": "Done"}, 404
